notifications: move queries file to common helpers
With time will reuse queries from the dashboard,
so this is just a first to share the work.
