+++ backend/kernelCI_app/queries/notifications.py
